# api/openapi.yaml
openapi: 3.0.3
info:
  title: Salu-Rec API
  description: フットサル試合管理・MVP選出アプリ API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: ローカル開発
security:
  - bearerAuth: []
...
paths:
...
  /api/events/{eventId}/rounds:
    post:
      operationId: createRound
      tags: [round-command]
      summary: ラウンド作成（チーム分け実行）
      parameters:
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoundRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoundResponse"
        "400":
          description: バリデーションエラー（メンバー不足、イベント未進行中など）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    get:
      operationId: listRounds
      tags: [round-query]
      summary: ラウンド一覧取得
      parameters:
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundListResponse"
  /api/events/{eventId}/rounds/{roundId}:
    get:
      operationId: getRoundDetail
      tags: [round-query]
      summary: ラウンド詳細取得
      parameters:
        - $ref: "#/components/parameters/EventId"
        - $ref: "#/components/parameters/RoundId"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoundDetailResponse"
        "404":
          description: ラウンドが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
      description: イベントID
...
    RoundId:
      name: roundId
      in: path
      required: true
      schema:
        type: string
      description: ラウンドID
...
  schemas:
...
    ApiErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: バリデーションエラーが発生しました

    # ─── Auth ────────────────────────────────────────────
...
    CreateRoundRequest:
      type: object
      required: [teamCount, memberIds]
      properties:
        teamCount:
          type: integer
          minimum: 2
          description: チーム数（2〜人数÷3）
        memberIds:
          type: array
          items:
            type: string
          minItems: 4
          description: チーム分け対象メンバーID一覧
    CreateRoundResponse:
      type: object
      required: [roundId, roundNumber, teamAssignment]
      properties:
        roundId:
          type: string
        roundNumber:
          type: integer
        teamAssignment:
          $ref: "#/components/schemas/TeamAssignment"
    TeamAssignment:
      type: object
      required: [teams]
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/TeamResponse"
    TeamResponse:
      type: object
      required: [name, memberIds]
      properties:
        name:
          type: string
        memberIds:
          type: array
          items:
            type: string
        captainId:
          type: string
          nullable: true
          description: キャプテンのメンバーID
    RoundListResponse:
      type: object
      required: [rounds]
      properties:
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/RoundListItemResponse"
    RoundListItemResponse:
      type: object
      required: [id, roundNumber, status, teamCount, matchCount]
      properties:
        id:
          type: string
        roundNumber:
          type: integer
        status:
          type: string
          enum: [InProgress, Finished]
        teamCount:
          type: integer
        matchCount:
          type: integer
    RoundDetailResponse:
      type: object
      required: [id, eventId, roundNumber, status, teamAssignment]
      properties:
        id:
          type: string
        eventId:
          type: string
        roundNumber:
          type: integer
        status:
          type: string
          enum: [InProgress, Finished]
        teamAssignment:
          $ref: "#/components/schemas/TeamAssignment"

    # ─── Match ───────────────────────────────────────────
...